move timeline service to modules
