+++ packages/client-web/src/modules/timeline/views/Timeline/index.tsx
@@ -1,7 +1,7 @@
 import styled from "styled-components";
 
 import TimelineEntry from "../TimelineEntry";
-import { Entry } from "../../../../services/timeline/types";
+import { Entry } from "../../types";
 
 const Wrapper = styled.div`
   display: flex;
